refactor: extract Question Controls, Filters and Results components
